Change the port env var for Heroku compatibility.
Rename the port env var from DEFAULT_LISTEN_PORT to plain PORT. The latter is used by Heroku and there doesn't seem to be any other way to specify the port as a different env var or to map the in-container port to the port that Heroku wants to expose--Heroku's Docker support is very limited in this regard.

I intentionally kept the changes to the minimal possible and didn't update comments, documentation, or var names or make the maintenance burden of this fork smaller. It's possible this change makes sense to be contributed upstream.