Feature add arm support and use chromium
This allows to build and run gotenberg on a raspbarry 4b.

We are using chromium as there is no arm build of chrome.

The tiny executable needs to be changed to the version that is
compatible with the correct architecture.

Also preparing for multi-architecture building of the docker images in
the publish script.

We removed the crosscompiling enviornment variables of go, as they are
preventing a compiling for arm and they will be set automatically.

fixes https://github.com/thecodingmachine/gotenberg/issues/213